backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
langchain==0.3.0
langchain-community==0.3.0
faiss-cpu==1.13.2
python-dotenv==1.0.0
tiktoken==0.8.0
pypdf==3.17.4
requests==2.31.0
regex
